microcluster/db: Consolidate all DB utility functions
