+++ microcluster/db/slices_test.go
@@ -8,7 +8,7 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/canonical/microcluster/v3/cluster/db"
+	"github.com/canonical/microcluster/v3/microcluster/db"
 )
 
 // Exercise possible failure modes.
